Created Desk and Order objects and connected input values to them.  Changed most Desk data types to int since we aren't giving decimal options.  Changed shipping speed to int too, instead of enum (still have surface material enum)

+++ Group1Desk/Desk.cs
@@ -10,14 +10,14 @@ namespace Group1Desk
 
     class Desk
     {
-        public double width { get; set; }
-        public double length { get; set; }
+        public int width { get; set; }
+        public int length { get; set; }
         public int drawers { get; set; }
         public SurfaceMaterial surfaceType { get; set; }
 
-        public double getSurfaceArea()
+        public int getSurfaceArea()
         {
-            double area;
+            int area;
             area = width * length;
             return area;
         }
